CLN: Split integer array tests (#32910)
